add RatingEngine refactoring and add auto, land and life policy services
