+++ ArdalisRating.Tests/RatingEngineRate.cs
@@ -1,3 +1,5 @@
+using ArdalisRating.Domain.Enums;
+using ArdalisRating.Services.Handlers;
 using ArdalisRating.Utils;
 using Newtonsoft.Json;
 using System.IO;
